Conflicts:
	src/main/java/es/uma/khaos/docking_service/model/ParameterSet.java
	src/main/java/es/uma/khaos/docking_service/resource/TaskResource.java
	src/main/java/es/uma/khaos/docking_service/service/DatabaseService.java

+++ src/main/java/es/uma/khaos/docking_service/model/response/TaskRunResponse.java
@@ -5,12 +5,11 @@ import java.util.ArrayList;
 public class TaskRunResponse {
 	
 	private int id;
-	private int finalBindingEnergy;
+	private float finalBindingEnergy;
 	private ArrayList<String> objectives;
 	private int execution_task_id;
 	
-	
-	public TaskRunResponse(int id, int finalBindingEnergy,
+	public TaskRunResponse(int id, float finalBindingEnergy,
 			ArrayList<String> objectives, int execution_task_id) {
 		super();
 		this.id = id;
@@ -19,22 +18,17 @@ public class TaskRunResponse {
 		this.execution_task_id = execution_task_id;
 	}
 
-
-
 	public int getId() {
 		return id;
 	}
 
-
-	public int getFinalBindingEnergy() {
+	public float getFinalBindingEnergy() {
 		return finalBindingEnergy;
 	}
-
 	
 	public ArrayList<String> getObjectives() {
 		return objectives;
 	}
-
 
 	public int getExecution_task_id() {
 		return execution_task_id;
